+++ src/libs/lib/auth.js
@@ -1,0 +1,11 @@
+let oscAuth = {};
+
+oscAuth.getLoginUrl = function(config) {
+  let returnTo = 'returnTo=' + encodeURIComponent(document.location.toString());
+  let loginUrl = config.loginUrl || '/oauth/login?{returnTo}';
+  loginUrl = loginUrl.replace(/{returnTo}/, returnTo);
+  console.log(loginUrl);
+  return loginUrl;
+}
+
+export default oscAuth;