+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.33.0
 * Fix error when clicking on new idea button in mobile view
 * Make votebuttons in idea-details configurable
+* Fix login redirects
 
 ## 0.32.0
 * Add new configuration options to reactions